# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Takeaway App API",
        "description": "API documentation for the Takeaway App",
        "contact": {
            "email": "linwenhao5@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customer/register": {
            "post": {
                "tags": [
                    "Customer Authentication"
                ],
                "summary": "Register a new customer",
                "operationId": "0e511cae3f862b478f6e1d3504955436",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Customer's name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Customer's email",
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "description": "Customer's password",
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "captcha": {
                                        "description": "Captcha code sent to the email",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "REGISTRATION_SUCCESS"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Customer registered successfully!"
                                        },
                                        "customer": {
                                            "description": "Customer details",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: Email already exists, invalid captcha, or other validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "Email already exists, invalid captcha, or validation failed",
                                            "type": "string",
                                            "example": "EMAIL_ALREADY_EXISTS"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The email has already been taken or Invalid captcha."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden: Registration is disabled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "REGISTRATION_DISABLED"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registration is currently disabled."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Invalid captcha",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "INVALID_CAPTCHA"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid captcha."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to register customer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "REGISTRATION_FAILED"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to register customer."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/login": {
            "post": {
                "tags": [
                    "Customer Authentication"
                ],
                "summary": "Login a customer",
                "operationId": "d759a66d56bf3c0487cb6feb118c7596",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Customer's email",
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "description": "Customer's password",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful!"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "LOGIN_SUCCESS"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "INVALID_CREDENTIALS"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to login."
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "LOGIN_FAILED"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/forgot-password": {
            "post": {
                "tags": [
                    "Customer Authentication"
                ],
                "summary": "Reset customer password using captcha",
                "operationId": "e78b51a7e55a04b3fb5ce031054ba67d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Customer's email",
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "captcha": {
                                        "description": "Captcha code sent to the email",
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password": {
                                        "description": "New password (min 8 chars)",
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password reset successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid captcha",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid captcha."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to reset password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to reset password."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/generate-captcha": {
            "post": {
                "tags": [
                    "Customer Authentication"
                ],
                "summary": "Generate a captcha code for email verification",
                "operationId": "d75b9a89e0ed2c294183f77e278f7080",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Customer's email",
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Captcha sent successfully or already sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "CAPTCHA_SENT_SUCCESS"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Captcha sent successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to generate captcha",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "CAPTCHA_GENERATE_FAILED"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to generate captcha."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token:",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}